Refactor NiceGUI implementation for ScribbleTrace: - Renamed main script to `scribbletrace-gui` for clarity. - Removed Gradio dependencies and migrated to NiceGUI framework. - Enhanced image processing pipeline with additional algorithms and improved image handling. - Introduced an image gallery for uploaded images with selection and deletion capabilities. - Updated histogram normalization and rendering functions for better performance. - Improved SVG generation and preview handling with dynamic sizing options. - Added GitHub link in the UI for easy access to the repository.
