+++ pyproject.toml
@@ -36,8 +36,6 @@ dependencies = [
 
 [project.optional-dependencies]
 gui = [
-    "gradio>=4.0",
-    "dearpygui>=1.11",
     "nicegui>=2.0",
 ]
 dev = [
@@ -47,9 +45,7 @@ dev = [
 
 [project.scripts]
 scribbletrace = "scribbletrace.cli:main"
-scribbletrace-gui = "scribbletrace.gui:main"
-scribbletrace-dpg = "scribbletrace.gui_dpg:main"
-scribbletrace-nicegui = "scribbletrace.gui_nicegui:main"
+scribbletrace-gui = "scribbletrace.gui_nicegui:main"
 
 [project.urls]
 Homepage = "https://github.com/kylberg/ScribbleTrace"
